Use sourceforge's load balancing / auto selection for downloads. Do not remove temporary files if unable to install binaries.
git-svn-id: https://svn.master.freemedsoftware.org/freemed/trunk@2880 6507a214-7cff-0310-bd63-e3842a560bc8

+++ scripts/build_gsdjvu.sh
@@ -26,10 +26,12 @@ VERSION="1.1"
 echo "build_gsdjvu.sh for gsdjvu version ${VERSION}"
 echo "by jeff@freemedsoftware.org"
 echo " "
+echo "Please note you must have root privileges to install this package."
+echo " "
 
 if [ ! -f "gsdjvu-${VERSION}.tar.gz" ]; then
 	echo -n " * Retrieving gsdjvu v${VERSION} ... "
-	wget -q -c http://easynews.dl.sourceforge.net/sourceforge/djvu/gsdjvu-${VERSION}.tar.gz 2>&1 > /dev/null
+	wget -q -c http://downloads.sourceforge.net/djvu/gsdjvu-${VERSION}.tar.gz 2>&1 > /dev/null
 	echo "[done]"
 else
 	echo " * Already have gsdjvu v${VERSION}"
@@ -37,7 +39,7 @@ fi
 
 if [ ! -f "ghostscript-8.15.tar.bz2" ]; then
 	echo -n " * Retrieving GPL ghostscript ... "
-	wget -q -c http://easynews.dl.sourceforge.net/sourceforge/ghostscript/ghostscript-8.15.tar.bz2 2>&1 > /dev/null
+	wget -q -c http://downloads.sourceforge.net/ghostscript/ghostscript-8.15.tar.bz2 2>&1 > /dev/null
 	echo "[done]"
 else
 	echo " * Already have GPL ghostscript package"
@@ -45,7 +47,7 @@ fi
 
 if [ ! -f "ghostscript-fonts-std-8.11.tar.gz" ]; then
 	echo -n " * Retrieving ghostscript fonts ... "
-	wget -q -c http://easynews.dl.sourceforge.net/sourceforge/ghostscript/ghostscript-fonts-std-8.11.tar.gz 2>&1 > /dev/null
+	wget -q -c http://downloads.sourceforge.net/ghostscript/ghostscript-fonts-std-8.11.tar.gz 2>&1 > /dev/null
 	echo "[done]"
 else
 	echo " * Already have ghostscript font package"
@@ -90,6 +92,11 @@ echo -n " * Moving into /usr/local/gsdjvu ... "
 mv -f gsdjvu-${VERSION}/BUILD/INST/gsdjvu /usr/local
 ( cd /usr/local/bin; ln -s /usr/local/gsdjvu/gsdjvu . )
 echo "[done]"
+
+if [ ! -f /usr/local/bin/gsdjvu ]; then
+	echo " * Was not able to install gsdjvu, please install binaries manually."
+	exit 1
+fi
 
 echo -n " * Cleaning up ... "
 rm -rf gsdjvu-${VERSION}
